Veri tabani postgrese cevrildi

+++ src/models.rs
@@ -1,25 +1,24 @@
-use crate::schema::sensor_veri;
-use chrono::NaiveDateTime;
+use crate::schema::sensor_data;
 use serde::{Deserialize, Serialize};
-
+use chrono::NaiveDateTime;
 #[derive(Queryable, Serialize, Deserialize, Debug, Clone)]
 pub struct SensorData {
     pub id: i32,
-    pub sicaklik: f32,
-    pub nem: f32,
+    pub sicaklik: f64,
+    pub nem: f64,
     pub timestamp: NaiveDateTime,
 }
 
 #[derive(Insertable)]
-#[table_name = "sensor_veri"]
+#[table_name = "sensor_data"]
 pub struct NewSensorData {
-    pub sicaklik: f32,
-    pub nem: f32,
+    pub sicaklik: f64,
+    pub nem: f64,
     pub timestamp: NaiveDateTime,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct SensorPostData {
-    pub sicaklik: f32,
-    pub nem: f32,
+    pub sicaklik: f64,
+    pub nem: f64,
 }
